[MIGR] sale_automatic_workflow: move code in the correct file
